+++ lib/SD_Operation_D/SD_Op_D.h
@@ -10,7 +10,7 @@ class SDOP{
     public:
         void            listAllDir(fs::FS &fs, const char * dirname, uint8_t levels);
         //uint8_t         listFile_pending(fs::FS &fs, const char * dirname, char *array_pending, uint8_t total_file_pending);
-        int16_t         countFile(fs::FS &fs, const char * dirname, uint8_t limit);
+        int16_t         countFile(fs::FS &fs, const char * dirname, uint16_t limit);
         void            createDir(fs::FS &fs, const char * path);
         void            removeDir(fs::FS &fs, const char * path);
         String          readFile(fs::FS &fs, const char * path);
